+++ HISTORY.md
@@ -8,6 +8,7 @@
 5. Implement last used text for each chat mode
 6. Change mode colours key to mode.value
 7. Click on pair assigns correct message 
+8. Clicking on history item assigns message
 
 
 ## Version 0.0.8 01 September 2026